Moved file injection.hpp
